webapp/requirements.txt
# This requirements file lists all third-party dependencies for this project.
#
# Run 'pip install -r requirements.txt -t lib/' to install these dependencies
# in `lib/` subdirectory.
#
# Note: The `lib` directory is added to `sys.path` by `appengine_config.py`.
Flask==0.10
oauth2client==1.5.1
httplib2==0.9.2
google-api-python-client==1.4.2
urllib3==1.14
certifi>=2015.11.20.1
python-dateutil
pyasn1
ruamel.yaml